Split the large server-actions file into three files with more specific responsibilities (deployments, instances, tasks)
