+++ src/management-system-v2/app/(dashboard)/[environmentId]/(automation)/executions/deployments-view.tsx
@@ -11,11 +11,11 @@ import { Process, ProcessMetadata } from '@/lib/data/process-schema';
 import { useEnvironment } from '@/components/auth-can';
 import { processHasChangesSinceLastVersion } from '@/lib/data/processes';
 import { useRouter } from 'next/navigation';
-import { deployProcess as serverDeployProcess } from '@/lib/engines/server-actions';
+import { deployProcess as serverDeployProcess } from '@/lib/executions/deployment-server-actions';
 import { wrapServerCall } from '@/lib/wrap-server-call';
 import { Engine } from '@/lib/engines/machines';
 import { userError } from '@/lib/user-error';
-import { removeDeployment as serverRemoveDeployment } from '@/lib/engines/server-actions';
+import { removeDeployment as serverRemoveDeployment } from '@/lib/executions/deployment-server-actions';
 import { useQueryClient } from '@tanstack/react-query';
 
 type InputItem = ProcessMetadata | (Folder & { type: 'folder' });
