+++ src/management-system-v2/app/api/private/[environmentId]/engine/resources/process/[definitionId]/instance/[instanceId]/file/[fileName]/route.ts
@@ -1,6 +1,6 @@
 import { UnauthorizedError } from '@/lib/ability/abilityHelper';
 import { auth } from '@/lib/auth';
-import { getFile } from '@/lib/engines/server-actions';
+import { getFile } from '@/lib/executions/instance-server-actions';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(
