+++ src/management-system-v2/lib/executions/deployment-server-actions.ts
@@ -10,72 +10,31 @@ import {
 import {
   deployProcess as _deployProcess,
   getDeployments as fetchDeployments,
-  getDeployment as fetchDeployment,
-  getProcessImageFromMachine,
   removeDeploymentFromMachines,
   changeDeploymentActivation as _changeDeploymentActivation,
   VersionInfo,
   InstanceInfo,
-} from './deployment';
-import { Engine } from './machines';
-import { savedEnginesToEngines } from './saved-engines-helpers';
+} from '../engines/deployment';
+import { Engine } from '../engines/machines';
+import { savedEnginesToEngines } from '../engines/saved-engines-helpers';
 import { getCurrentEnvironment, getCurrentUser } from '@/components/auth';
-import {
-  getMachineIfAvailable,
-  getAvailableAdminMachines,
-  getAllAvailableMachines,
-} from '@/lib/data/engines';
+import { getAvailableAdminMachines, getAllAvailableMachines } from '@/lib/data/engines';
 import { asyncFilter, asyncMap, asyncForEach, AsyncArray } from '../helpers/javascriptHelpers';
 
 import db from '@/lib/data/db';
 
-import {
-  activateUserTask,
-  completeTasklistEntryOnMachine,
-  getTaskListFromMachine,
-  setTasklistEntryVariableValuesOnMachine,
-  setTasklistEntryMilestoneValuesOnMachine,
-  addOwnerToTaskListEntryOnMachine,
-} from './tasklist';
+import { getTaskListFromMachine } from '../engines/tasklist';
 import { MapNestedType, Prettify, truthyFilter } from '../typescript-utils';
 import {
-  inlineUserTaskData,
   getCorrectVariableState,
   getCorrectMilestoneState,
   inlineScript,
-  getGlobalVariables,
 } from '@proceed/user-task-helper';
 
-import { getUserTaskById, updateUserTask } from '../data/user-tasks';
-import {
-  getFileFromMachine,
-  pauseInstanceOnMachine,
-  resumeInstanceOnMachine,
-  startInstanceOnMachine,
-  stopInstanceOnMachine,
-  submitFileToMachine,
-  updateVariablesOnMachine,
-} from './instances';
-import {
-  getProcessIds,
-  getStartFormFileNameMapping,
-  getVariablesFromElementById,
-} from '@proceed/bpmn-helper';
+import { getProcessIds, getVariablesFromElementById } from '@proceed/bpmn-helper';
 import { Variable } from '@proceed/bpmn-helper/src/getters';
 import { addDeployment, updateDeployment, getProcessDeployments } from '../data/deployment';
-import {
-  addInstance,
-  getInstance,
-  getInstance as getStoredInstance,
-  updateInstance,
-} from '../data/instance';
-import {
-  getProcessBPMN,
-  getProcessHtmlFormHTML,
-  getProcessImage as getStoredProcessImage,
-} from '../data/processes';
-import { getDataObject, isErrorResponse } from '@/app/api/spaces/[spaceId]/data/helper';
-import { getInstanceFile, saveInstanceArtifact } from '../data/file-manager-facade';
+import { getProcessBPMN, getProcessHtmlFormHTML } from '../data/processes';
 import { revalidateTag } from 'next/cache';
 
 export async function deployProcess(
@@ -86,6 +45,11 @@ export async function deployProcess(
   _forceEngine?: Engine | 'PROCEED',
 ) {
   try {
+    const { ability } = await getCurrentEnvironment(spaceId);
+
+    if (!ability.can('create', 'Execution'))
+      return userError('Invalid Permissions', UserErrorType.PermissionError);
+
     // TODO: manage permissions for deploying a process
     const { userId } = await getCurrentUser();
 
@@ -111,6 +75,7 @@ export async function deployProcess(
 
     // find all currently available deployments
     const processAlreadyDeployedInfo = alreadyDeployed
+      .filter((d) => !d.deleted)
       .map((deployment) => ({
         ...deployment,
         machines: deployment.machineIds
@@ -268,643 +233,6 @@ export async function changeDeploymentActivation(
     const message = getErrorMessage(e);
     return userError(message);
   }
-}
-
-export async function getGlobalVariablesForHTML(
-  spaceId: string,
-  initiatorId: string,
-  html: string,
-) {
-  return await getGlobalVariables(html, async (varPath) => {
-    let segments = varPath.split('.');
-
-    let userId: string | undefined;
-
-    if (segments[0] === '@process-initiator') {
-      userId = initiatorId;
-    } else if (segments[0] === '@worker' || !segments[0].startsWith('@')) {
-      ({ userId } = await getCurrentUser());
-    } else if (segments[0] !== '@organization') {
-      throw new UserFacingError(
-        `Invalid selector for global data access in user task html. (${segments[0]})`,
-      );
-    }
-
-    if (segments[0].startsWith('@')) segments = segments.slice(1);
-
-    const result = await getDataObject(spaceId, segments.join('.'), userId);
-
-    if (isErrorResponse(result)) {
-      throw new UserFacingError(await result.data.text());
-    }
-
-    return result.data?.value;
-  });
-}
-
-export async function getTasklistEntryHTML(spaceId: string, userTaskId: string) {
-  try {
-    const storedUserTask = await getUserTaskById(spaceId, userTaskId);
-
-    if (!storedUserTask || isUserErrorResponse(storedUserTask)) {
-      throw new Error('Failed to get stored user task data.');
-    }
-
-    let { initialVariables, variableChanges, milestones, milestonesChanges, html } = storedUserTask;
-
-    if (storedUserTask.state === 'READY') {
-      let activated = true;
-      if (storedUserTask.instanceID && storedUserTask.machineId) {
-        try {
-          const engine = await getMachineIfAvailable(spaceId, storedUserTask.machineId);
-          if (engine) {
-            await activateUserTask(
-              engine,
-              storedUserTask.instanceID,
-              storedUserTask.taskId,
-              storedUserTask.startTime.getTime(),
-            );
-          } else {
-            activated = false;
-          }
-        } catch (err) {
-          activated = false;
-        }
-      }
-
-      if (activated) await updateUserTask(spaceId, userTaskId, { state: 'ACTIVE' });
-    }
-
-    variableChanges = { ...initialVariables, ...(variableChanges || {}) };
-
-    if (milestonesChanges) {
-      milestones = (milestones || []).map((milestone) => ({
-        ...milestone,
-        value: milestonesChanges![milestone.id] ?? milestone.value,
-      }));
-    }
-
-    // maps relative urls used to get resources on the engine to the MS api to allow them to work here as well
-    function mapResourceUrls(variables: Record<string, any>) {
-      if (!variables) return variables;
-
-      return Object.fromEntries(
-        Object.entries(variables).map(([key, value]) => {
-          const [_, instanceId] = userTaskId.split('|');
-          const [definitionId] = instanceId.split('-_');
-
-          if (
-            typeof value === 'string' &&
-            value.includes(`resources/process/${definitionId}/instance/${instanceId}/file/`)
-          ) {
-            return [key, `api/private/${spaceId}/engine/` + value];
-          }
-
-          return [key, value];
-        }),
-      );
-    }
-
-    let globalVars: Record<string, any> = {};
-
-    if (storedUserTask.instanceID) {
-      const instance = await getStoredInstance(spaceId, storedUserTask.instanceID);
-      if (isUserErrorResponse(instance)) return instance;
-      if (!instance) throw new Error('Cannot retrieve the instance initiator information.');
-      if (!instance.initiatorId) throw new Error('Missing initiator information');
-
-      globalVars = await getGlobalVariablesForHTML(spaceId, instance.initiatorId, html);
-
-      variableChanges = mapResourceUrls(variableChanges);
-    }
-
-    variableChanges = { ...variableChanges, ...globalVars };
-
-    return inlineUserTaskData(html, variableChanges, milestones);
-  } catch (e) {
-    const message = getErrorMessage(e);
-    return userError(message);
-  }
-}
-
-export async function addOwnerToTaskListEntry(
-  spaceId: string,
-  userTaskId: string,
-  ownerId: string,
-) {
-  try {
-    const storedUserTask = await getUserTaskById(spaceId, userTaskId);
-
-    if (!storedUserTask || 'error' in storedUserTask) {
-      throw new Error('Failed to get stored user task data.');
-    }
-
-    let { actualOwner } = storedUserTask;
-
-    if (!actualOwner.some((owner) => owner.id === ownerId)) {
-      await updateUserTask(spaceId, userTaskId, {
-        actualOwner: [...actualOwner.map((aO) => aO.id), ownerId],
-      });
-
-      if (storedUserTask.instanceID) {
-        const { machineId } = storedUserTask;
-        const engine = machineId && (await getMachineIfAvailable(spaceId, machineId));
-
-        if (!engine) {
-          return userError('Could not find the engine this user task is running on.');
-        }
-
-        const [taskId, instanceId] = userTaskId.split('|');
-
-        return await addOwnerToTaskListEntryOnMachine(engine, instanceId, taskId, ownerId);
-      }
-    }
-
-    return {};
-  } catch (e) {
-    const message = getErrorMessage(e);
-    return userError(message);
-  }
-}
-
-export async function setTasklistEntryVariableValues(
-  spaceId: string,
-  userTaskId: string,
-  variables: { [key: string]: any },
-) {
-  try {
-    const storedUserTask = await getUserTaskById(spaceId, userTaskId);
-
-    if (!storedUserTask || 'error' in storedUserTask) {
-      throw new Error('Failed to get stored user task data.');
-    }
-
-    await updateUserTask(spaceId, userTaskId, {
-      variableChanges: { ...storedUserTask.variableChanges, ...variables },
-    });
-
-    if (storedUserTask.instanceID) {
-      const { machineId } = storedUserTask;
-      const engine = machineId && (await getMachineIfAvailable(spaceId, machineId));
-
-      if (!engine) {
-        return userError('Could not find the engine this user task is running on.');
-      }
-
-      const [taskId, instanceId] = userTaskId.split('|');
-
-      await setTasklistEntryVariableValuesOnMachine(engine, instanceId, taskId, variables);
-    }
-  } catch (e) {
-    const message = getErrorMessage(e);
-    return userError(message);
-  }
-}
-
-export async function setTasklistMilestoneValues(
-  spaceId: string,
-  userTaskId: string,
-  milestones: { [key: string]: any },
-) {
-  try {
-    const storedUserTask = await getUserTaskById(spaceId, userTaskId);
-
-    if (!storedUserTask || 'error' in storedUserTask) {
-      throw new Error('Failed to get stored user task data.');
-    }
-
-    await updateUserTask(spaceId, userTaskId, {
-      milestonesChanges: { ...storedUserTask.milestonesChanges, ...milestones },
-    });
-
-    if (storedUserTask.instanceID) {
-      const { machineId } = storedUserTask;
-      const engine = machineId && (await getMachineIfAvailable(spaceId, machineId));
-
-      if (!engine) {
-        return userError('Could not find the engine this user task is running on.');
-      }
-
-      const [taskId, instanceId] = userTaskId.split('|');
-
-      await setTasklistEntryMilestoneValuesOnMachine(engine, instanceId, taskId, milestones);
-    }
-  } catch (e) {
-    const message = getErrorMessage(e);
-    return userError(message);
-  }
-}
-
-export async function completeTasklistEntry(
-  spaceId: string,
-  userTaskId: string,
-  variables: { [key: string]: any },
-) {
-  try {
-    const storedUserTask = await getUserTaskById(spaceId, userTaskId);
-
-    if (!storedUserTask || 'error' in storedUserTask) {
-      throw new Error('Failed to get stored user task data.');
-    }
-
-    const { variableChanges, milestonesChanges } = storedUserTask;
-
-    if (storedUserTask.instanceID) {
-      const { machineId } = storedUserTask;
-      const engine = machineId && (await getMachineIfAvailable(spaceId, machineId));
-
-      if (!engine) {
-        return userError('Could not find the engine this user task is running on.');
-      }
-
-      const [taskId, instanceId] = userTaskId.split('|');
-
-      // push the values from the database to the engine so the instance state is correctly updated
-      // when the user task is completed as the next step
-      await setTasklistEntryVariableValuesOnMachine(
-        engine,
-        instanceId,
-        taskId,
-        variableChanges || {},
-      );
-      await setTasklistEntryMilestoneValuesOnMachine(
-        engine,
-        instanceId,
-        taskId,
-        milestonesChanges || {},
-      );
-
-      await completeTasklistEntryOnMachine(engine, instanceId, taskId, variables);
-    }
-
-    await updateUserTask(spaceId, userTaskId, {
-      variableChanges: { ...variableChanges, ...variables },
-      state: 'COMPLETED',
-      endTime: !storedUserTask.instanceID ? new Date() : undefined,
-    });
-  } catch (e) {
-    const message = getErrorMessage(e);
-    return userError(message);
-  }
-}
-
-export async function updateVariables(
-  spaceId: string,
-  definitionId: string,
-  instanceId: string,
-  variables: Record<string, any>,
-) {
-  try {
-    const instance = await getStoredInstance(spaceId, instanceId);
-
-    if (isUserErrorResponse(instance)) return instance;
-
-    if (!instance) return userError('Could not find the instance to change.');
-
-    // find the engine the instance is running on
-    const engines = await AsyncArray.from(getAllAvailableMachines(spaceId)).filter((engine) => {
-      return instance.machineIds.includes(engine.id);
-    });
-
-    if (!engines.length) return userError('Could not reach the engines the instance is running on');
-
-    await asyncForEach(
-      engines,
-      async (engine) => await updateVariablesOnMachine(definitionId, instanceId, engine, variables),
-    );
-
-    await new Promise((res) => setTimeout(res, 1000));
-
-    // TODO: handle that we need to merge data if the instance exists on multiple machines
-    const newData = await fetchDeployment(engines[0], definitionId, 'instances');
-
-    const newInstanceData = newData.instances.find((i) => i.processInstanceId === instanceId);
-
-    await updateInstance(spaceId, instanceId, { state: newInstanceData });
-  } catch (err) {
-    const message = getErrorMessage(err);
-    return userError(message);
-  }
-}
-
-export async function submitFile(spaceId: string, userTaskId: string, formData: FormData) {
-  try {
-    const storedUserTask = await getUserTaskById(spaceId, userTaskId);
-
-    if (!storedUserTask || 'error' in storedUserTask) {
-      throw new Error('Failed to get stored user task data.');
-    }
-
-    if (!storedUserTask.instanceID) {
-      return userError('We cannot save files for locally created user tasks');
-    }
-
-    const { machineId } = storedUserTask;
-    const engine = machineId && (await getMachineIfAvailable(spaceId, machineId));
-
-    if (!engine) {
-      return userError('Could not find the engine this user task is running on.');
-    }
-
-    const file = formData.get('file') as File;
-
-    const fileName = file.name;
-    const fileType = file.type;
-
-    const [_, instanceId] = userTaskId.split('|');
-    const [definitionId] = instanceId.split('-_');
-
-    if (!engine) throw new Error('Could not find the engine to submit the file to');
-
-    const res = await submitFileToMachine(
-      definitionId,
-      instanceId,
-      engine,
-      fileName,
-      fileType,
-      Array.from(new Uint8Array(await file.arrayBuffer())),
-    );
-
-    await saveInstanceArtifact(
-      spaceId,
-      instanceId,
-      fileName,
-      file.type,
-      Buffer.from(await file.arrayBuffer()),
-    );
-
-    return res;
-  } catch (err) {
-    const message = getErrorMessage(err);
-    return userError(message);
-  }
-}
-
-export async function getFile(
-  spaceId: string,
-  definitionId: string,
-  instanceId: string,
-  fileName: string,
-) {
-  const instance = await getStoredInstance(spaceId, instanceId);
-
-  if (isUserErrorResponse(instance)) return instance;
-
-  if (!instance) {
-    throw new UserFacingError('Unknown instance.');
-  }
-
-  const savedFile = await getInstanceFile(instanceId, fileName);
-  if (savedFile) {
-    return savedFile;
-  }
-
-  try {
-    // find the engine the instance is running on
-    const engines = await AsyncArray.from(getAllAvailableMachines(spaceId)).filter((engine) => {
-      return instance.machineIds.includes(engine.id);
-    });
-
-    if (!engines.length) {
-      throw new UserFacingError('Failed to find the engine the instance is running on!');
-    }
-
-    const file = await getFileFromMachine(definitionId, instanceId, fileName, engines[0]);
-
-    try {
-      const res = await saveInstanceArtifact(
-        spaceId,
-        instanceId,
-        fileName,
-        file.type,
-        Buffer.from(await file.arrayBuffer()),
-      );
-
-      if (res.presignedUrl) {
-        // TODO: handle cloud storage case
-      }
-    } catch (err) {
-      console.error(`Failed to cache instance file ${fileName}.`);
-    }
-
-    return file;
-  } catch (err) {
-    const message = getErrorMessage(err);
-    return userError(message);
-  }
-}
-
-export async function getProcessImage(spaceId: string, instanceId: string, fileName: string) {
-  try {
-    const instance = await getInstance(spaceId, instanceId);
-    if (isUserErrorResponse(instance)) return instance;
-    if (!instance) return userError('Unknown instance');
-
-    // try to get the image locally
-    const image = await getStoredProcessImage(instance.state.processId, fileName, spaceId);
-
-    if (!isUserErrorResponse(image)) {
-      return new Blob([image]);
-    }
-
-    // if the image is not found locally try to get it from the engines that are executing the
-    // process
-    const engines = await AsyncArray.from(getAllAvailableMachines(spaceId)).filter((engine) => {
-      return instance.machineIds.includes(engine.id);
-    });
-
-    if (!engines.length) throw new Error('Failed to get an engine the process was deployed to!');
-
-    return await getProcessImageFromMachine(engines[0], instance.state.processId, fileName);
-  } catch (err) {
-    const message = getErrorMessage(err);
-    return userError(message);
-  }
-}
-
-export async function getProcessStartForm(
-  spaceId: string,
-  definitionId: string,
-  versionId: string,
-) {
-  try {
-    const bpmn = await getProcessBPMN(definitionId, spaceId, versionId);
-    const [startForm] = Object.values(await getStartFormFileNameMapping(bpmn)).filter(truthyFilter);
-    if (!startForm) return '';
-
-    return getProcessHtmlFormHTML(definitionId, startForm, spaceId);
-  } catch (e) {
-    const message = getErrorMessage(e);
-    return userError(message);
-  }
-}
-
-export async function startInstance(
-  spaceId: string,
-  definitionId: string,
-  versionId: string,
-  variables: { [key: string]: any } = {},
-) {
-  const machines = await getAllAvailableMachines(spaceId);
-  const machineMap = machines.reduce(
-    (acc, curr) => {
-      acc[curr.id] = curr;
-      return acc;
-    },
-    {} as Record<string, Engine>,
-  );
-  const deployments = await getProcessDeployments(spaceId, definitionId);
-  const { userId } = await getCurrentUser();
-
-  if (isUserErrorResponse(deployments)) return deployments;
-
-  const versionDeployments = deployments.filter((d) => {
-    return d.versionId === versionId && d.machineIds.some((id) => !!machineMap[id]);
-  });
-
-  // TODO: automatically deploy the version if possible
-  if (!versionDeployments.length) return userError('This process version is not deployed.');
-
-  for (const deployment of deployments) {
-    const availableMachines = deployment.machineIds.map((id) => machineMap[id]).filter((m) => !!m);
-
-    for (const machine of availableMachines) {
-      // TODO: maybe have the engine return the complete instance information instead of just the id on
-      // instance creation
-      const result = await startInstanceOnMachine(definitionId, versionId, machine, variables, {
-        processInitiator: userId,
-        spaceIdOfProcessInitiator: spaceId,
-      });
-
-      if (isUserErrorResponse(result)) continue;
-
-      const engineDeploymentInfo = await fetchDeployment(machine, definitionId);
-      const instance = engineDeploymentInfo.instances.find((i) => i.processInstanceId === result);
-
-      if (!instance) return userError('Failed to fetch the newly created instance');
-
-      await addInstance(spaceId, {
-        id: result,
-        versionId,
-        deploymentId: versionDeployments[0].id,
-        machineIds: [machine.id],
-        initiatorId: userId,
-        state: instance,
-      });
-
-      return result;
-    }
-  }
-
-  return userError('Failed to start an instance.');
-}
-
-const activeStates = ['PAUSED', 'RUNNING', 'READY', 'DEPLOYMENT-WAITING', 'WAITING'];
-async function changeInstanceState(
-  spaceId: string,
-  definitionId: string,
-  instanceId: string,
-  stateValidator: (state: InstanceInfo['instanceState']) => boolean,
-  stateChangeFunction: typeof resumeInstanceOnMachine,
-) {
-  const instance = await getStoredInstance(spaceId, instanceId);
-
-  if (isUserErrorResponse(instance)) return instance;
-  if (!instance) return userError('Unknown instance!', UserErrorType.NotFoundError);
-
-  if (!instance.machineIds.length) return userError('The instance is not being executed anymore.');
-
-  try {
-    // TODO: how do we handle this correctly if some engines are reachable but others aren't?
-    const machines = await getAllAvailableMachines(spaceId);
-    const machineMap = machines.reduce(
-      (acc, curr) => {
-        acc[curr.id] = curr;
-        return acc;
-      },
-      {} as Record<string, Engine>,
-    );
-
-    const instanceMachines = await asyncMap(instance.machineIds, async (id) => {
-      const machine = machineMap[id];
-      if (!machine) return false;
-
-      try {
-        const deployment = await fetchDeployment(machine, definitionId);
-
-        const instance = deployment.instances.find(
-          (instance) => instance.processInstanceId === instanceId,
-        );
-
-        if (!instance) return 'Instance not found';
-
-        const hasState = !stateValidator(instance.instanceState);
-
-        if (hasState) return false;
-
-        return machine;
-      } catch (err) {
-        return false;
-      }
-    }).filter((res) => !!res);
-
-    console.log(instanceMachines);
-    if (instanceMachines.some((res) => typeof res === 'string')) {
-      return userError('Instance information was lost from one of the executing engines.');
-    }
-
-    await asyncForEach(instanceMachines as Engine[], async (machine) => {
-      await stateChangeFunction(definitionId, instanceId, machine);
-
-      // TODO: handle this better (the engine should only return after the state change has
-      // completed
-      await new Promise((res) => setTimeout(res, 1000));
-      // TODO: actually handle that the instance might exist on multiple engines
-      const newDeploymentData = await fetchDeployment(machine, definitionId);
-      const newInstanceData = newDeploymentData.instances.find(
-        (instance) => instance.processInstanceId === instanceId,
-      );
-      await updateInstance(spaceId, instanceId, { state: newInstanceData });
-    });
-  } catch (e) {
-    console.log(e);
-    const message = getErrorMessage(e);
-    return userError(message);
-  }
-}
-
-export async function resumeInstance(spaceId: string, definitionId: string, instanceId: string) {
-  // TODO: manage permissions for starting an instance
-  return await changeInstanceState(
-    spaceId,
-    definitionId,
-    instanceId,
-    (tokenStates) => tokenStates.some((tokenState) => tokenState === 'PAUSED'),
-    resumeInstanceOnMachine,
-  );
-}
-
-export async function pauseInstance(spaceId: string, definitionId: string, instanceId: string) {
-  // TODO: manage permissions for starting an instance
-  return await changeInstanceState(
-    spaceId,
-    definitionId,
-    instanceId,
-    (tokenStates) =>
-      tokenStates.some((state) => activeStates.includes(state) && state !== 'PAUSED'),
-    pauseInstanceOnMachine,
-  );
-}
-
-export async function stopInstance(spaceId: string, definitionId: string, instanceId: string) {
-  // TODO: manage permissions for starting an instance
-  return await changeInstanceState(
-    spaceId,
-    definitionId,
-    instanceId,
-    (tokenStates) => tokenStates.some((state) => activeStates.includes(state)),
-    stopInstanceOnMachine,
-  );
 }
 
 // all connected users are sharing this information
